Added mechanism to stop docker_init.sh if there's no docker running

+++ docker_init.sh
@@ -1,4 +1,18 @@
-docker stop $(docker ps -q -a)
-docker rm $(docker ps -q -a)
+DOCKER_PROCESS=$(docker ps | head -1)
+
+if [ ${DOCKER_PROCESS:0:9} == "CONTAINER" ]
+then
+    echo "Docker is up!"
+
+    echo "docker stop all"
+    docker stop $(docker ps -q -a)
+
+    echo "docker rm all"
+    docker rm $(docker ps -q -a)
+
+    echo "docker build backend"
+    docker build -t backend .
+fi
+
+###############################
 #docker rmi $(docker images -q)
-docker build -t backend .
